refactor: 커넥션 객체를 싱글톤으로 관리하도록 변경
