+++ src/main/java/dao/ConnectionProvider.java
@@ -1,0 +1,23 @@
+package dao;
+
+import static dao.DatabaseConfig.*;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class ConnectionProvider {
+
+    private static Connection connection;
+
+    public static Connection getConnection() {
+        if (connection == null) {
+            try {
+                connection = DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
+            } catch (final SQLException e) {
+                e.printStackTrace();
+            }
+        }
+        return connection;
+    }
+}
